minor fixes and updates

+++ javascript/pages/index.js
@@ -39,7 +39,7 @@ const preview_code_traitsHTML =
 `use std::math::{PI}
 
 trait Shape {
-    fn area(self): int
+    fn area(self): f32
 }
 
 struct Rectangle {
@@ -48,7 +48,7 @@ struct Rectangle {
 }
 
 impl Shape for Rectangle {
-    fn area(self): int {
+    fn area(self): f32 {
         ret self.width * self.height
     }
 }
@@ -58,7 +58,7 @@ struct Circle {
 }
 
 impl Shape for Circle {
-    fn area(self): int {
+    fn area(self): f32 {
         ret PI * self.r * self.r
     }
 }
